+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project uses [Semantic Versioning](http://semver.org/).
 
+# [2.0.0] - 2020-7-10
+### Added
+ - Octopod learners can now use multiple loss functions and do multi-label in addition to multi-class tasks
+### Changed
+ - loss functions and metrics are specified via a dictionary of tasks and corresponding loss functions 
 
 # [1.0.0] - 2020-7-09
 ### Changed